requirements.txt
agora_token_builder==1.0.0
asgiref==3.8.1
Django==5.1.3
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.2
